fix GetGroupInfo() panic.
